server_requirements.txt
flask
flask-cors
torch
torchvision
timm
numpy
pillow
firebase-admin
